Вынесла форму профиля в отдельный компонент
