+++ src/pages/profile/profile.tsx
@@ -1,26 +1,23 @@
 import profileStyles from './profile.module.css';
-import {
-  Button,
-  Input,
-  PasswordInput
-} from '@ya.praktikum/react-developer-burger-ui-components';
-import React, { FormEvent, useEffect, useRef, useState } from 'react';
+import { Button } from '@ya.praktikum/react-developer-burger-ui-components';
+import React, { useEffect, useState } from 'react';
 import { Layout } from '../../components/layout/layout';
 import { useDispatch, useSelector } from 'react-redux';
-import { RegisterUser } from '../../shared/models/register-user.type';
-import {
-  fetchUserThunk,
-  fetchUserUpdatingThunk
-} from '../../services/actions/user';
+import { fetchUserThunk } from '../../services/actions/user';
 import { UnknownAction } from 'redux';
 import { fetchLogoutThunk } from '../../services/actions/login';
 import { useNavigate } from 'react-router-dom';
 import { Routes as RouteName } from '../../shared/consts/routes';
+import { FeedCardImage } from '../../components/feed-card-image/feed-card-image';
+import { ProfileForm } from '../../components/profile-form/profile-form';
 import { Logout } from '../../shared/models/store/logout.type';
-import { useForm } from '../../shared/hooks/use-form';
+
+enum Parts {
+  Profile = 'profile',
+  History = 'history'
+}
 
 type ProfileSelector = {
-  user: RegisterUser;
   login: Logout;
 };
 
@@ -28,43 +25,12 @@ export function ProfilePage() {
   const navigate = useNavigate();
   const dispatch = useDispatch();
   const useProfilePageSelector = useSelector.withTypes<ProfileSelector>();
-  const user = useProfilePageSelector(state => state.user);
   const logout = useProfilePageSelector(state => state.login.logout);
-  const [values, handleChange, setCertainValue] = useForm<
-    Required<RegisterUser>
-  >({
-    email: '',
-    password: '',
-    name: ''
-  });
-  const [showButtons, setShowButtons] = useState(false);
-  const nameRef = useRef<HTMLInputElement>(null);
-  const loginRef = useRef<HTMLInputElement>(null);
-
-  useEffect(() => {
-    if (user && Object.keys(user).length) {
-      if (user?.name) {
-        setCertainValue('name', user.name);
-      }
-      setCertainValue('email', user.email);
-    }
-  }, [user]);
+  const [currentPart, setCurrentPart] = useState<Parts>(Parts.Profile);
 
   useEffect(() => {
     dispatch(fetchUserThunk() as unknown as UnknownAction);
   }, []);
-
-  useEffect(() => {
-    if (
-      user.name !== nameRef.current?.value ||
-      user.email !== loginRef.current?.value ||
-      values.password.length
-    ) {
-      setShowButtons(true);
-    } else {
-      setShowButtons(false);
-    }
-  }, [user, nameRef.current?.value, loginRef.current?.value, values.password]);
 
   const onLogout = () => {
     dispatch(fetchLogoutThunk() as unknown as UnknownAction);
@@ -76,30 +42,31 @@ export function ProfilePage() {
     }
   }, [logout]);
 
-  const onSaveProfile = (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    dispatch(
-      fetchUserUpdatingThunk({
-        ...values
-      }) as unknown as UnknownAction
-    );
-  };
-
-  const onCancelEdit = () => {
-    setCertainValue('name', user.name);
-    setCertainValue('email', user.email);
-    setShowButtons(false);
-  };
-
   return (
     <Layout>
       <div className={`${profileStyles.grid} pt-30`}>
         <div className={`${profileStyles.col} ${profileStyles.textWidth}`}>
-          <p className='text text_type_main-large pb-6'>Профиль</p>
-          <p className='text text_type_main-large text_color_inactive pb-6'>
-            История заказов
-          </p>
-          <div className={profileStyles.logoutButton}>
+          <div className={profileStyles.button}>
+            <Button
+              htmlType='button'
+              type='secondary'
+              size='medium'
+              onClick={() => setCurrentPart(Parts.Profile)}
+            >
+              <p className='text text_type_main-large pb-6'>Профиль</p>
+            </Button>
+          </div>
+          <div className={profileStyles.button}>
+            <Button
+              htmlType='button'
+              type='secondary'
+              size='medium'
+              onClick={() => setCurrentPart(Parts.History)}
+            >
+              <p className='text text_type_main-large pb-6'>История заказов</p>
+            </Button>
+          </div>
+          <div className={profileStyles.button}>
             <Button
               htmlType='button'
               type='secondary'
@@ -116,73 +83,11 @@ export function ProfilePage() {
           </div>
         </div>
 
-        <form className='pl-15' onSubmit={onSaveProfile}>
-          <div className={profileStyles.col}>
-            <div className='pb-6'>
-              <Input
-                ref={nameRef}
-                type='text'
-                placeholder='Имя'
-                onChange={handleChange}
-                value={values.name}
-                name={'name'}
-                error={false}
-                errorText={'Ошибка'}
-                size={'default'}
-                extraClass='ml-1'
-                icon='EditIcon'
-                autoComplete='name'
-                onPointerEnterCapture={() => ({})}
-                onPointerLeaveCapture={() => ({})}
-              />
-            </div>
-            <div className='pb-6'>
-              <Input
-                ref={loginRef}
-                type='text'
-                placeholder='Логин'
-                onChange={handleChange}
-                value={values.email}
-                name={'login'}
-                error={false}
-                icon='EditIcon'
-                errorText={'Ошибка'}
-                size={'default'}
-                extraClass='ml-1'
-                autoComplete='email'
-                onPointerEnterCapture={() => ({})}
-                onPointerLeaveCapture={() => ({})}
-              />
-            </div>
-            <PasswordInput
-              placeholder='Пароль'
-              onChange={handleChange}
-              value={values.password}
-              name={'password'}
-              icon='EditIcon'
-              errorText={'Ошибка'}
-              size={'default'}
-              extraClass='ml-1'
-              autoComplete='current-password'
-            />
-          </div>
-
-          {showButtons ? (
-            <div className={`${profileStyles.saveButton} pt-10`}>
-              <Button
-                htmlType='button'
-                type='secondary'
-                size='medium'
-                onClick={onCancelEdit}
-              >
-                Отмена
-              </Button>
-              <Button htmlType='submit' type='primary' size='medium'>
-                Сохранить
-              </Button>
-            </div>
-          ) : null}
-        </form>
+        {currentPart === Parts.Profile ? (
+          <ProfileForm />
+        ) : (
+          <FeedCardImage index={1} />
+        )}
       </div>
     </Layout>
   );
